# backend/openapi/api.json
{"openapi":"3.1.0","info":{"title":"delhi-anpr-api","version":"0.1.0"},"paths":{"/healthz":{"get":{"summary":"Healthz","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/cameras":{"get":{"summary":"Cameras","description":"Camera registry with geometry \u2014 feeds map pins and trajectory joins.","operationId":"cameras_api_cameras_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/vehicles/search":{"get":{"summary":"Search","operationId":"search_api_vehicles_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","title":"Q"}},{"name":"fuzzy","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Fuzzy"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/vehicles/{canonical}/trajectory":{"get":{"summary":"Trajectory","operationId":"trajectory_api_vehicles__canonical__trajectory_get","parameters":[{"name":"canonical","in":"path","required":true,"schema":{"type":"string","title":"Canonical"}},{"name":"from_ts","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From Ts"}},{"name":"to_ts","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To Ts"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/vehicles/{canonical}/clones":{"get":{"summary":"Clones","operationId":"clones_api_vehicles__canonical__clones_get","parameters":[{"name":"canonical","in":"path","required":true,"schema":{"type":"string","title":"Canonical"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/heatmap":{"get":{"summary":"Heatmap","operationId":"heatmap_api_analytics_heatmap_get","parameters":[{"name":"bucket","in":"query","required":false,"schema":{"type":"string","default":"5m","enum":["1m","5m","15m","1h","6h","24h"],"title":"Bucket","description":"Re-bin stride for the 1-min camera_rollups (date_bin aggregation). Unknown values return 400."}},{"name":"from_ts","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From Ts"}},{"name":"to_ts","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To Ts"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/congestion":{"get":{"summary":"Congestion","operationId":"congestion_api_analytics_congestion_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/analytics/od":{"get":{"summary":"Od","operationId":"od_api_analytics_od_get","parameters":[{"name":"window","in":"query","required":false,"schema":{"type":"string","default":"30m","title":"Window"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/cameras/health":{"get":{"summary":"Health Cams","operationId":"health_cams_api_analytics_cameras_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/sweeps/zone":{"get":{"summary":"Sweep Zone","description":"Area sweep: recent reads from cameras within radius_km of (lat, lng).\nPowers the Area Sweep panel (P2: was mock-only). Rows share the\ntrajectory shape (+crop_url) so the frontend reuses its live builder.","operationId":"sweep_zone_api_sweeps_zone_get","parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","title":"Lat"}},{"name":"lng","in":"query","required":true,"schema":{"type":"number","title":"Lng"}},{"name":"radius_km","in":"query","required":false,"schema":{"type":"number","maximum":20,"exclusiveMinimum":0,"default":1.0,"title":"Radius Km"}},{"name":"from_ts","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From Ts"}},{"name":"to_ts","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To Ts"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"exclusiveMinimum":0,"default":200,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sweeps/window":{"get":{"summary":"Sweep Window","description":"Time-window sweep: grid-wide summary + top plates for the window.\nPowers the Time Window panel (P2: was mock-only).","operationId":"sweep_window_api_sweeps_window_get","parameters":[{"name":"from_ts","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From Ts"}},{"name":"to_ts","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To Ts"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"exclusiveMinimum":0,"default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/alerts":{"get":{"summary":"Alerts","operationId":"alerts_api_alerts_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}},{"name":"include_acknowledged","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Include Acknowledged"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/blacklist":{"get":{"summary":"List Blacklist","description":"Wanted-plate registry \u2014 feeds the admin panel (P1: was POST-only).","operationId":"list_blacklist_api_blacklist_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Add Blacklist","operationId":"add_blacklist_api_blacklist_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Body"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/crops/{name}":{"get":{"summary":"Get Crop","description":"Serve a persisted plate crop. Empty until a source posts crop_b64\n(the sim sends none today) \u2014 the gallery renders empty, never faked.","operationId":"get_crop_api_crops__name__get","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/blacklist/{plate}":{"delete":{"summary":"Remove Blacklist","operationId":"remove_blacklist_api_blacklist__plate__delete","parameters":[{"name":"plate","in":"path","required":true,"schema":{"type":"string","title":"Plate"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/alerts/{alert_id}/ack":{"post":{"summary":"Ack Alert","description":"Acknowledge an alert \u2014 drives the inbox ack button (P1: column existed, unwired).","operationId":"ack_alert_api_alerts__alert_id__ack_post","parameters":[{"name":"alert_id","in":"path","required":true,"schema":{"type":"string","title":"Alert Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/accuracy/report":{"get":{"summary":"Accuracy Report","operationId":"accuracy_report_api_accuracy_report_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}